+++ docs/releasenotes/releasenotes.rst
@@ -15,6 +15,7 @@ Enhancements
 * custom.js: New hook which will be called before first Dashticz render: ``beforeFirstRenderHook()``. See :ref:`beforefirstrenderhook`
 * Styling: Add correct data-id to a Dashticz block consisting of multiple blocks. See :ref:`grouped_devices`
 * World Air Quality Index: Support for the WAQI widgets have been added. See :ref:`waqi`
+* publictransport: ``'drgl'`` as new provider, which will give bus and train departures in the Netherlands. See :ref:`drgl`
 
 Fixes
 ~~~~~~